Minor Adjustments and move to account release version

+++ Sources/SpeziFirebaseAccount/Account Services/FirebaseAccountService.swift
@@ -26,16 +26,12 @@ protocol FirebaseAccountService: AnyActor, AccountService {
     /// - Parameter context: The global firebase context
     func configure(with context: FirebaseContext) async
 
-    /// This method is called once the account for the given user was removed.
-    ///
-    /// This allows for additional cleanup tasks to be performed.
-    /// - Parameter userId: The userId which was removed, or nil if we couldn't retrieve the last user.
-    func handleAccountRemoval(userId: String?) async
-
     /// This method is called to re-authenticate the current user credentials.
-    /// - Parameters:
-    ///   - user: The User instance.
-    func reauthenticateUser(user: User) async throws
+    ///
+    /// - Parameter user: The user instance to reauthenticate.
+    /// - Returns: `true` if authentication was successful, `false` if authentication was cancelled by the user.
+    /// - Throws: If authentication failed.
+    func reauthenticateUser(user: User) async throws -> Bool // TODO: properly type?
 }
 
 
@@ -72,12 +68,18 @@ extension FirebaseAccountService {
         }
 
         try await context.dispatchFirebaseAuthAction(on: self) {
-            try await reauthenticateUser(user: currentUser) // delete requires a recent sign in
+            let valid = try await reauthenticateUser(user: currentUser) // delete requires a recent sign in
+            guard valid else {
+                Self.logger.debug("Re-authentication was cancelled. Not deleting the account.")
+                return // cancelled
+            }
+
             try await currentUser.delete()
             Self.logger.debug("delete() for user.")
         }
     }
 
+    // swiftlint:disable:next cyclomatic_complexity
     func updateAccountDetails(_ modifications: AccountModifications) async throws {
         guard let currentUser = Auth.auth().currentUser else {
             if await account.signedIn {
@@ -88,12 +90,16 @@ extension FirebaseAccountService {
 
         var changes = false
 
-        // if we modify sensitive credentials and require a recent login
-        if modifications.modifiedDetails.storage[UserIdKey.self] != nil || modifications.modifiedDetails.password != nil {
-            try await reauthenticateUser(user: currentUser)
-        }
-
         do {
+            // if we modify sensitive credentials and require a recent login
+            if modifications.modifiedDetails.storage[UserIdKey.self] != nil || modifications.modifiedDetails.password != nil {
+                let valid = try await reauthenticateUser(user: currentUser)
+                guard valid else {
+                    Self.logger.debug("Re-authentication was cancelled. Not deleting the account.")
+                    return // got cancelled!
+                }
+            }
+
             if let userId = modifications.modifiedDetails.storage[UserIdKey.self] {
                 Self.logger.debug("updateEmail(to:) for user.")
                 try await currentUser.updateEmail(to: userId)
@@ -103,10 +109,6 @@ extension FirebaseAccountService {
             if let password = modifications.modifiedDetails.password {
                 Self.logger.debug("updatePassword(to:) for user.")
                 try await currentUser.updatePassword(to: password)
-
-                if let userId = currentUser.email { // make sure we save the new password
-                    await context.persistCurrentCredentials(userId: userId, password: password, server: StorageKeys.emailPasswordCredentials)
-                }
             }
 
             if let name = modifications.modifiedDetails.name {
@@ -123,8 +125,10 @@ extension FirebaseAccountService {
                 try await context.notifyUserSignIn(user: currentUser, for: self)
             }
         } catch let error as NSError {
+            Self.logger.error("Received NSError on firebase dispatch: \(error)")
             throw FirebaseAccountError(authErrorCode: AuthErrorCode(_nsError: error))
         } catch {
+            Self.logger.error("Received error on firebase dispatch: \(error)")
             throw FirebaseAccountError.unknown(.internalError)
         }
     }
